+++ eduVPN.Views/Converters/RenewVisibilityConverter.cs
@@ -38,7 +38,11 @@ namespace eduVPN.Converters
             return
                 values[1] is DateTimeOffset from && from != DateTimeOffset.MinValue &&
                 values[2] is DateTimeOffset to && to != DateTimeOffset.MaxValue &&
+#if DEBUG
+                (now - from).TotalMinutes > 1 &&
+#else
                 (now - from).TotalMinutes > 30 &&
+#endif
                 (to - now).TotalHours < 24 ?
                 Visibility.Visible : Visibility.Collapsed;
         }
@@ -56,6 +60,6 @@ namespace eduVPN.Converters
             throw new NotImplementedException();
         }
 
-        #endregion
+#endregion
     }
 }