+++ CHANGES.md
@@ -2,6 +2,7 @@
 
 ## [Unreleased](https://github.com/Amebis/eduVPN/compare/3.4...HEAD)
 
+- Session Renewal fixed
 - OpenVPN updated to 2.5.9-20230427
     - openvpn 2.5.9 ea4ce681d9008f277706f4d90f2648ae043cbb2e
 - libsodium updated to 1.0.18-20230427